+++ scheduler/server.c
@@ -1,7 +1,7 @@
 /*
  * Server start/stop routines for the CUPS scheduler.
  *
- * Copyright 2007-2018 by Apple Inc.
+ * Copyright 2007-2019 by Apple Inc.
  * Copyright 1997-2006 by Easy Software Products, all rights reserved.
  *
  * These coded instructions, statements, and computer programs are the
@@ -175,6 +175,15 @@ cupsdStopServer(void)
 
   cupsdDestroyProfile(DefaultProfile);
   DefaultProfile = NULL;
+
+ /*
+  * Expire subscriptions and clean out old jobs...
+  */
+
+  cupsdExpireSubscriptions(NULL, NULL);
+
+  if (JobHistoryUpdate)
+    cupsdCleanJobs();
 
  /*
   * Write out any dirty files...