+++ scheduler/subscriptions.c
@@ -1,7 +1,7 @@
 /*
  * Subscription routines for the CUPS scheduler.
  *
- * Copyright 2007-2014 by Apple Inc.
+ * Copyright 2007-2019 by Apple Inc.
  * Copyright 1997-2007 by Easy Software Products, all rights reserved.
  *
  * These coded instructions, statements, and computer programs are the
@@ -651,6 +651,9 @@ cupsdExpireSubscriptions(
   int			update;		/* Update subscriptions.conf? */
   time_t		curtime;	/* Current time */
 
+
+  if (cupsArrayCount(Subscriptions) == 0)
+    return;
 
   curtime = time(NULL);
   update  = 0;